AllocatedBuffer now stores the buffers size and minAlignment.
VertexBuffer now offers a helper function to obtain the capacity of elements

Buffer now offers two public functions to obtain the capacity in bytes and the minAlignment the buffer were created with.

Allocator now stores minimum Uniform and Storage alignments, which it uses when it allocates these buffers.

Added first draft of GPU types which are to be shared between glsl and cpp