Allocators create_storage_buffer no longer takes a VkBufferCreateInfo as argument. Instead it only takes a VkDeviceSize.
Added a helper function to Allocator, to_staging(), that takes a contiguous range, creates a staging buffer, writes the range content to it before returning the staging buffer itself.

Fixed a bug in Buffer.hpp that caused it to not write the full byte range.

Added DRAW_FLAGS to gpu_types - to aid the shader understand what vertex data is available.